# src/openapi.yaml
openapi: 3.0.0
info:
  title: Seal Issuing Service
  version: '1.0'
  description: API to issue seals for confidential data and track their usage and revelation
servers:
  - url: 'http://localhost:8080/api/v1'
paths:
  /seals:
    post:
      summary: Create a new seal
      operationId: createSeal
      x-eov-operation-handler: handlers/seals
      tags:
        - Seal Management
      requestBody:
        description: Optional parameters for creating the seal
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SealRequest'
      responses:
        201:
          description: Creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seal'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /seals/{id}:
    get:
      summary: Get a seal by its id
      description: |
        Returns all information about the seal including sensitive data such as key and salt.
        CAUTION: Although this endpoints returns the key, it does not increase the revelation_token_max_use counter.
      operationId: getSeal
      x-eov-operation-handler: handlers/seals
      tags:
        - Seal Management
      security:
        - OwnershipTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seal'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Update a seal by its id
      operationId: updateSeal
      x-eov-operation-handler: handlers/seals
      tags:
        - Seal Management
      security:
        - OwnershipTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Update for seal
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SealUpdate'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealUpdate'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Deletes a seal including all tracking capabilities
      operationId: deleteSeal
      x-eov-operation-handler: handlers/seals
      tags:
        - Seal Management
      security:
        - OwnershipTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Deletion successful
        default:
          $ref: '#/components/responses/ErrorResponse'
  /seals/{id}/events:
    get:
      summary: Get all events of a seal
      operationId: getSealEvents
      x-eov-operation-handler: handlers/seals
      tags:
        - Seal Management
      security:
        - OwnershipTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SealEvent'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /seals/{id}/revelations:
    post:
      summary: Reveal a given seal
      operationId: revealSeal
      x-eov-operation-handler: handlers/seals
      tags:
        - Seal Revelation
      security:
        - RevelationTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealRevelation'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    SealRequest:
      title: SealRequest
      type: object
      properties:
        revelation_token_max_use:
          description: |
            Maximum count of revelation token use
            Set to 0 if token can be used without limit
          type: number
        not_before:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    SealUpdate:
      title: SealUpdate
      type: object
      properties:
        status:
          type: string
          enum:
            - revoked
    SealRevelation:
      title: SealRevelation
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        key:
          type: string
          readOnly: true
        salt:
          type: string
          readOnly: true
    Seal:
      title: Seal
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the seal
          format: uuid
          readOnly: true
        key:
          type: string
          description: Secret key, use this for symmetric encryption
          readOnly: true
        salt:
          type: string
          description: |
            Salt value for protecting against dictionary attacks
            Attach this to the plaintext value before encryption
          readOnly: true
        revelation_token:
          type: string
          description: |
            Token to reveal the secret in case of decryption
            Must be passed to the recipient
          readOnly: true
        revelation_token_max_use:
          type: number
          description: |
            Maximum count of revelation token use
            Set to 0 if token can be used without limit
        ownership_token:
          type: string
          description: |
            Token to manage the seal after issuing, e. g for deletion or revocation
            Must NOT be passed to the recipient
          readOnly: true
        status:
          type: string
          description: Status of the seal
          enum:
            - unrevealed
            - revealed
            - revoked
            - inactive
          readOnly: true
        not_before:
          type: string
          description: Seal cannot be revealed before this timestamp
          format: date-time
        expires_at:
          type: string
          description: Seal cannot be revealed after this timestamp
          format: date-time
    SealEvent:
      title: SealEvent
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        eventName:
          type: string
          enum:
            - create
            - revoke
            - reveal
            - delete
    Error:
      title: Error
      type: object
      required:
        - message
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
  securitySchemes:
    RevelationTokenAuth:
      type: http
      scheme: bearer
    OwnershipTokenAuth:
      type: http
      scheme: bearer